requirements.txt
Flask==2.1.3
Flask-Cors==3.0.10
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-core==2.3.1
google-cloud-speech==2.15.0
google-cloud-storage==2.4.0
google-cloud-translate==3.8.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
websocket-client==1.3.3
Werkzeug==2.1.2
zipp==3.8.1
